Added new game menu and updated file save functions
When choosing a save file the program checks if the file is new or not. Then if the file is new will open a newGame menu, which leads into the locker room menu. However when updating the save file to be read as not a new save the data does not go into the correct line on the txt file, will be fixed next update hopefully.