Added create name menu
When a new save is started the enter name menu is shown with a functional on screen keyboard to enter you name which is updated and displayed onto the screen. When the name is entered the save file is updated with the users name and the user is taken to the locker room menu.