Added more start of game functions
When a team is selected a round robin schedule is created and uploaded to the save file. The player and team class objects are created and the teams are filled with players, then the player is swapped with an NPC on the team they picked. Need to update schedule to include players and there stats, not just each each and the teams