Fixed bug when changing between different save files
Everytime a save was opened, new or old, that players name was added to the list of players making the list too long causing loops to go out of bounds. Now the current name is checked if it is already on the list before it is added. Also the current name is removed when going to the main menu from the locker room.

(Potential Bug: Player makes their name something like "save 2" or "buzz" which would break the search for player name function. Potential fix: add a " " to the end of every user name so it can't match another string in the save file.)