Bug Fix
Bug where game would crash when switching between profiles. This was caused from the users name not getting added when loading a save so when you switch saves it would get deleted, then not added again so when you go to start menu a second time the playerNames list would go out of bounds. This is now fixed.

Also, started player stats page. Most of the UI is implemented, and some variables are initialized.