Player Goals are read from save file when cache is created
When the player caches are being created the save file is read and retrieves the number of goals per game for each player.

This brings up a major issue I had though of before. With just this implemented the loading time as sky rocketed to several minutes, and with additional stats to be loaded in still this calls for a MAJOR REWORK of the file save system.