Reworked save file loading and saving and cache loading and saving
Reformatted the save file structure to allow for better cache loading time. Now cache loads player data in ~28 seconds, while before it took several minutes just to load the player names, teams, and career goals.

Implementation is not finished. Stats page is no longer accessible as it causes the game to crash and needs to be fixed.