Updated Player Stats Page & Bug Fix
Fixed bug where the caches were not being loaded when opening an already created save. This was due to not resetting the player and team class list to [], they where getting reset to [0] which was part of the old system. This is now fixed.

The player stats page has been updated to show the player goals again. Now there are also two buttons to switch between season stats and career stats, so you can currently see the different stats in the goals column.

Next update will work on finishing the player stats page by loading in the rest of the stats per player. Then functions will be made to sort the date by type in the stats table.