Updated GUI for the game menu
GUI now has stat numbers with headers and player names. Some arrows to add to stats have been added, and team season records show. The rest of the arrows need to be added with functionality to add and subtract each stat number. Then the functionality of the simulation and confirm buttons needs to be added. Also a GUI element showing the game score total for each team should be added.