Update to backend playoff logic
Now when the final game of the season is over the logic for the playoffs begin, the save files have been updated to hold, read and write data for the playoff schedule, and the team objects now have playoff records being stored. The logic to generate a new playoff schedule and simulate the playoff games has been implemented for the first 3 weeks of the playoffs. The other weeks of playoffs logic will be very similar but has not been implemented yet. The play next menu has been updated to show if the player is on a bye week, or has been eliminated from playoffs. Some GUI elements need to be adjusted to look better still.